+++ changelog/CHANGELOG.md
@@ -4,12 +4,13 @@
 
 2019/09/12
 
-* 加入命令行的-v,-h,command的help等操作
+* 加入命令行的-v,-h,commands的help等操作
 * 整理修改startapp的command参数
 * 整理修改buildapp的command参数
 * 整理修改buildumd的command参数
 * 整理修改buildpackage的command参数
 * 去掉了--type参数
+* 更新了clean-webpack-plugin插件为最新版本
 
 # 1.0.7